main.dart de-cluttered (#110)
